components: move wizard to new Wizard directory
The CreateImageUpload wizard should be in a wizard directory instead of
the modal directory.
